Set some cache expiry times on our static assets
I've set a 'relatively' short time-out on our own assets. It's long
enough that you should be able to complete an application, but short
enough that you'll get a fresh copy of everything 'after lunch'. The
govuk assets are a lot less sensitive, so I've given them a few hours so
they'll definitely be wiped 'by tomorrow'.

Issue: Scottish-Natural-Heritage/Licensing#301