Amend beforeEach in testing
This is a temporary work around for this postcode lookup issue we experienced in Ravens and now in SFO.
Arron suggested it's because I'm not running the start:mock which gives the hard coded post codes.
If I run server:mock, I have an issue with the session cookie not saving so I can't progress through the pages.
While Arron is looking at a fix today, I've edited this beforeEach step through so I can work on the tests.
I'll revert it once we've found a solution.

Issue: Scottish-Natural-Heritage/Licensing#1670